Extract/expand failing combinatorial tests.
Add throw-on-enumeration source.
